chore: move token logic into tokens package
