+++ server/src/types.rs
@@ -1,28 +1,19 @@
 use std::fmt;
 use std::sync::Arc;
 use std::{
-    future::{ready, Ready},
     hash::{Hash, Hasher},
     str::FromStr,
 };
 
-use crate::cli::EdgeMode;
 use crate::error::EdgeError;
-use actix_web::web::Data;
-use actix_web::{
-    dev::Payload,
-    http::header::{EntityTag, HeaderValue},
-    web::Json,
-    FromRequest, HttpRequest,
-};
+use actix_web::{http::header::EntityTag, web::Json};
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use shadow_rs::shadow;
 use unleash_types::client_features::ClientFeatures;
 use unleash_types::client_metrics::{ClientApplication, ClientMetricsEnv};
-use unleash_types::frontend::EvaluatedToggle;
-use unleash_yggdrasil::{Context, EngineState};
+use unleash_yggdrasil::EngineState;
 use utoipa::ToSchema;
 
 pub type EdgeJsonResult<T> = Result<Json<T>, EdgeError>;
@@ -104,127 +95,6 @@ impl Hash for EdgeToken {
     }
 }
 
-impl EdgeToken {
-    pub fn no_project_or_environment(s: &str) -> Self {
-        EdgeToken {
-            token: s.into(),
-            token_type: None,
-            environment: None,
-            projects: vec![],
-            status: TokenValidationStatus::default(),
-        }
-    }
-
-    pub fn subsumes(&self, other: &EdgeToken) -> bool {
-        return self.token_type == other.token_type
-            && self.environment == other.environment
-            && (self.projects.contains(&"*".into())
-                || (self.projects.len() >= other.projects.len()
-                    && other.projects.iter().all(|p| self.projects.contains(p))));
-    }
-}
-
-impl FromRequest for EdgeToken {
-    type Error = EdgeError;
-    type Future = Ready<EdgeResult<Self>>;
-
-    fn from_request(req: &HttpRequest, _payload: &mut Payload) -> Self::Future {
-        let value = req.headers().get("Authorization");
-        if let Some(data_mode) = req.app_data::<Data<EdgeMode>>() {
-            let mode = data_mode.clone().into_inner();
-            let key = match *mode {
-                EdgeMode::Offline(_) => match value {
-                    Some(v) => match v.to_str() {
-                        Ok(value) => Ok(EdgeToken::offline_token(value)),
-                        Err(_) => Err(EdgeError::AuthorizationDenied),
-                    },
-                    None => Err(EdgeError::AuthorizationDenied),
-                },
-                EdgeMode::Edge(_) => match value {
-                    Some(v) => EdgeToken::try_from(v.clone()),
-                    None => Err(EdgeError::AuthorizationDenied),
-                },
-            };
-            ready(key)
-        } else {
-            let key = match value {
-                Some(v) => EdgeToken::try_from(v.clone()),
-                None => Err(EdgeError::AuthorizationDenied),
-            };
-            ready(key)
-        }
-    }
-}
-
-impl TryFrom<HeaderValue> for EdgeToken {
-    type Error = EdgeError;
-
-    fn try_from(value: HeaderValue) -> Result<Self, Self::Error> {
-        value
-            .to_str()
-            .map_err(|_| EdgeError::AuthorizationDenied)
-            .and_then(EdgeToken::from_str)
-    }
-}
-
-impl TryFrom<String> for EdgeToken {
-    type Error = EdgeError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        EdgeToken::from_str(value.as_str())
-    }
-}
-
-impl FromStr for EdgeToken {
-    type Err = EdgeError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if s.contains(':') && s.contains('.') {
-            let token_parts: Vec<String> = s.split(':').take(2).map(|s| s.to_string()).collect();
-            let token_projects = if let Some(projects) = token_parts.get(0) {
-                if projects == "[]" {
-                    vec![]
-                } else {
-                    vec![projects.clone()]
-                }
-            } else {
-                return Err(EdgeError::TokenParseError);
-            };
-            if let Some(env_and_key) = token_parts.get(1) {
-                let e_a_k: Vec<String> = env_and_key
-                    .split('.')
-                    .take(2)
-                    .map(|s| s.to_string())
-                    .collect();
-                if e_a_k.len() != 2 {
-                    return Err(EdgeError::TokenParseError);
-                }
-                Ok(EdgeToken {
-                    environment: e_a_k.get(0).cloned(),
-                    projects: token_projects,
-                    token_type: None,
-                    token: s.into(),
-                    status: TokenValidationStatus::Unknown,
-                })
-            } else {
-                Err(EdgeError::TokenParseError)
-            }
-        } else {
-            Err(EdgeError::TokenParseError)
-        }
-    }
-}
-
-impl EdgeToken {
-    pub fn offline_token(s: &str) -> Self {
-        let mut token = EdgeToken::try_from(s.to_string())
-            .ok()
-            .unwrap_or_else(|| EdgeToken::no_project_or_environment(s));
-        token.status = TokenValidationStatus::Validated;
-        token
-    }
-}
-
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, utoipa::ToSchema)]
 pub struct TokenStrings {
     pub tokens: Vec<String>,
@@ -232,27 +102,6 @@ pub struct TokenStrings {
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, utoipa::ToSchema)]
 pub struct ValidatedTokens {
     pub tokens: Vec<EdgeToken>,
-}
-
-#[async_trait]
-pub trait FeatureSource {
-    async fn get_client_features(&self, token: &EdgeToken) -> EdgeResult<ClientFeatures>;
-
-    async fn resolve_computed_toggles(
-        &self,
-        edge_token: &EdgeToken,
-        context: &Context,
-        include_disabled: bool,
-    ) -> EdgeResult<Vec<EvaluatedToggle>>;
-}
-
-#[async_trait]
-pub trait TokenSource {
-    async fn get_tokens(&self) -> EdgeResult<Vec<EdgeToken>>;
-    async fn get_valid_tokens(&self) -> EdgeResult<Vec<EdgeToken>>;
-    async fn get_token(&self, secret: String) -> EdgeResult<Option<EdgeToken>>;
-    async fn filter_valid_tokens(&self, tokens: Vec<String>) -> EdgeResult<Vec<EdgeToken>>;
-    async fn get_tokens_due_for_refresh(&self) -> EdgeResult<Vec<TokenRefresh>>;
 }
 
 #[derive(Clone, Deserialize, Serialize)]
